requirements.txt
Django
requests
sentence-transformers
faiss-cpu
numpy